requirements.txt
bigtree==0.16.2
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
docutils==0.20.1
durations-nlp==1.0.1
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-examples==2.3.0
Kivy-Garden==0.1.5
multisort==0.1.5
oauthlib==3.2.2
pillow==10.2.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.1
pypiwin32==223
python-dateutil==2.8.2
pywin32==306
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tqdm==4.66.2
uritemplate==4.1.1
urllib3==2.2.0
